requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
bcrypt==4.3.0
binaryornot==0.4.4
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
cookiecutter==2.6.0
cryptography==45.0.4
fastapi==0.115.12
flatbuffers==20181003210633
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.27.6
flet-desktop-light==0.28.3
flet-web==0.28.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.2
onnxruntime==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
paramiko==3.5.1
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyMatting==1.1.14
PyNaCl==1.5.0
pypng==0.20220715.0
pysmb==1.2.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
rembg==2.0.66
repath==0.9.0
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
RPi.GPIO==0.7.1
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
text-unidecode==1.3
tifffile==2025.6.1
toml==0.10.2
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchdog==4.0.2
watchfiles==1.0.5
websockets==15.0.1
